backend/app/middleware/fraud.py
import logging
from typing import BinaryIO, Union
from PIL import Image
import io
import os

# Configure logging
logger = logging.getLogger(__name__)

class FraudDetectionError(Exception):
    """Custom exception for fraud detection errors"""
    pass

def validate_image_metadata(file: BinaryIO) -> bool:
    """
    Validate basic image metadata and format
    
    Args:
        file: File-like object containing the image
        
    Returns:
        bool: True if valid, False if invalid
    """
    try:
        # List of allowed MIME types
        ALLOWED_MIME_TYPES = ['image/jpeg', 'image/png']
        
        # Check content type
        content_type = getattr(file, 'content_type', None)
        if not content_type or content_type not in ALLOWED_MIME_TYPES:
            logger.warning(f"Invalid content type: {content_type}")
            return False
            
        # Try to open and validate the image
        try:
            image = Image.open(file)
            image.verify()  # Verify it's actually an image
            
            # Additional metadata checks
            if not image.format or image.format.upper() not in ['JPEG', 'PNG']:
                logger.warning(f"Invalid image format: {image.format}")
                return False
                
            # Check reasonable dimensions (e.g., not too small or too large)
            width, height = image.size
            if width < 100 or height < 100 or width > 10000 or height > 10000:
                logger.warning(f"Invalid image dimensions: {width}x{height}")
                return False
                
            return True
            
        except Exception as e:
            logger.error(f"Error validating image: {str(e)}")
            return False
            
    except Exception as e:
        logger.error(f"Error in metadata validation: {str(e)}")
        return False

def detect_fraud_upload(file: BinaryIO) -> Union[bool, None]:
    """
    Detect potential fraud in uploaded images using metadata analysis and AI-based detection
    
    Args:
        file: File-like object containing the image to check
        
    Returns:
        bool: True if fraud is detected, False if no fraud detected, None if analysis fails
        
    Raises:
        FraudDetectionError: If there's an error during fraud detection
    """
    try:
        # First validate basic metadata
        if not validate_image_metadata(file):
            logger.warning("Metadata validation failed")
            return True  # Consider invalid metadata as potential fraud
            
        # Reset file pointer for AI analysis
        file.seek(0)
        
        try:
            # TODO: Initialize your fraud detection model here
            # This is a placeholder - replace with actual AI model implementation
            fraud_probability = 0.0  # Replace with: fraud_detection_model.predict(file)
            
            # Log the fraud probability for monitoring
            logger.info(f"Fraud probability: {fraud_probability}")
            
            # Return True if probability exceeds threshold
            return fraud_probability > 0.7
            
        except Exception as e:
            logger.error(f"Error in AI fraud detection: {str(e)}")
            raise FraudDetectionError(f"AI fraud detection failed: {str(e)}")
            
    except Exception as e:
        logger.error(f"Unexpected error in fraud detection: {str(e)}")
        raise FraudDetectionError(f"Fraud detection failed: {str(e)}")
        
    finally:
        # Always reset file pointer
        try:
            file.seek(0)
        except:
            pass 