backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-jwt-extended==4.6.0
pymongo==4.6.2
python-dotenv==1.0.1
bson==0.5.10
Pillow==10.2.0  # For image processing in fraud detection